+++ index.ts
@@ -1,13 +1,13 @@
-export * from './lib/core/exports';
+export * from "./lib/modules/plugin/exports";
 
-export * from './lib/modules/asset/exports';
+export * from "./lib/modules/asset/exports";
 
-export * from './lib/modules/device/exports';
+export * from "./lib/modules/device/exports";
 
-export * from './lib/modules/measure/exports';
+export * from "./lib/modules/measure/exports";
 
-export * from './lib/modules/decoder/exports';
+export * from "./lib/modules/decoder/exports";
 
-export * from './lib/modules/shared/exports';
+export * from "./lib/modules/shared/exports";
 
-export * from './lib/modules/model/exports';
+export * from "./lib/modules/model/exports";
