--- lib/core/exports.ts
@@ -1,2 +1,0 @@
-export * from "./DeviceManagerConfiguration";
-export * from "./DeviceManagerPlugin";
