--- lib/core/index.ts
@@ -1,4 +1,0 @@
-export * from "./DeviceManagerConfiguration";
-export * from "./DeviceManagerEngine";
-export * from "./DeviceManagerPlugin";
-export * from "./InternalCollection";
