+++ lib/modules/decoder/PayloadsController.ts
@@ -1,7 +1,6 @@
 import { ControllerDefinition, KuzzleRequest } from "kuzzle";
 
-import { DecodersRegister } from "../../core/registers/DecodersRegister";
-
+import { DecodersRegister } from "./DecodersRegister";
 import { PayloadService } from "./PayloadService";
 import { ApiPayloadReceiveUnkownResult } from "./types/PayloadApi";
