+++ lib/modules/model/index.ts
@@ -5,3 +5,4 @@ export * from "./types/ModelDefinition";
 export * from "./types/ModelEvents";
 export * from "./ModelModule";
 export * from "./ModelSerializer";
+export * from "./ModelsRegister";
