+++ lib/modules/plugin/DeviceManagerEngine.ts
@@ -3,23 +3,20 @@ import { Backend, InternalError, Plugin } from "kuzzle";
 import { JSONObject } from "kuzzle-sdk";
 import { AbstractEngine, ConfigManager } from "kuzzle-plugin-commons";
 
-import { assetsMappings, assetsHistoryMappings } from "../modules/asset";
+import { assetsMappings, assetsHistoryMappings } from "../asset";
 import {
   AssetModelContent,
   DeviceModelContent,
   MeasureModelContent,
-} from "../modules/model";
-import {
-  getEmbeddedMeasureMappings,
-  measuresMappings,
-} from "../modules/measure";
-import { devicesMappings } from "../modules/device";
-import { onAsk } from "../modules/shared";
-import { NamedMeasures } from "../modules/decoder";
+} from "../model";
+import { getEmbeddedMeasureMappings, measuresMappings } from "../measure";
+import { devicesMappings } from "../device";
+import { onAsk } from "../shared";
+import { NamedMeasures } from "../decoder";
 
-import { DeviceManagerConfiguration } from "./DeviceManagerConfiguration";
+import { DeviceManagerConfiguration } from "./types/DeviceManagerConfiguration";
 import { DeviceManagerPlugin } from "./DeviceManagerPlugin";
-import { InternalCollection } from "./InternalCollection";
+import { InternalCollection } from "./types/InternalCollection";
 
 const digitalTwinMappings = {
   asset: assetsMappings,
