+++ lib/modules/plugin/exports.ts
@@ -1,0 +1,2 @@
+export * from "./types/DeviceManagerConfiguration";
+export * from "./DeviceManagerPlugin";
