+++ lib/modules/plugin/types/DeviceManagerConfiguration.ts
