+++ lib/modules/plugin/types/InternalCollection.ts
